Reuse override logic
